+++ src/MigrationAssistant/App.xaml.cs
