+++ src/MigrationAssistant/AssemblyInfo.cs
