+++ src/MigrationAssistant/MigrationAssistant.csproj
